tmp-chat-service/requirements.txt
annotated-types==0.7.0
anyio==4.5.2
async-timeout==5.0.1
bcrypt==3.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.15
distro==1.9.0
dnspython==2.6.1
exceptiongroup==1.2.2
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib-resources==6.4.5
jiter==0.9.0
limits==3.13.0
load-dotenv==0.1.0
openai==1.68.2
packaging==24.2
pycparser==2.22
pydantic==2.10.4
pydantic-core==2.27.2
PyJWT==2.4.0
pymongo==4.10.1
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
tqdm==4.67.1
typing-extensions==4.12.2
urllib3==2.2.3
uvicorn==0.33.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
wrapt==1.17.0
zipp==3.20.2
